core/requirements.txt
# general modules
django>=4.2,<4.3
python-decouple
pillow
djangorestframework
joblib
scikit-learn

# third-party modules
markdown     
django-filter
drf-yasg[validation]
djangorestframework-simplejwt[crypto]
djoser
django-cors-headers

# deployment modules
gunicorn
psycopg2-binary

# refactoring and testing modules
black
flake8



